requirements.txt
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
docstring_parser==0.16
google-api-core==2.19.1
google-auth==2.33.0
google-cloud-core==2.4.1
google-cloud-storage==2.18.2
google-crc32c==1.5.0
google-resumable-media==2.7.2
googleapis-common-protos==1.63.2
grpcio==1.65.4
idna==3.7
joblib==1.4.2
kfp==2.2.0
kfp-pipeline-spec==0.2.2
kfp-server-api==2.0.5
kubeflow-katib==0.16.0
kubernetes==26.1.0
numpy==2.0.1
nvidia-nccl-cu12==2.22.3
oauthlib==3.2.2
pandas==2.2.2
proto-plus==1.24.0
protobuf==3.20.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==0.10.1
rsa==4.9
scikit-learn==1.5.1
scipy==1.14.0
six==1.16.0
tabulate==0.9.0
threadpoolctl==3.5.0
tzdata==2024.1
urllib3==1.26.19
websocket-client==1.8.0
xgboost==2.1.1
